requirements.txt
kafka-python==2.0.2
numpy==1.19.4
py4j==0.10.9
pyspark==3.0.1
pyspark-stubs==3.0.0.post1
